started adding unit tests
Made unit test files
