+++ test/test_getFile.py
@@ -1,0 +1,10 @@
+import unittest
+from features import SFTP
+from unittest.mock import patch
+
+
+class UnitTests(unittest.TestCase):
+    # needs put and delete to tested easily
+    def test_getFile(self):
+        # print(self.sftp)
+        pass
