+++ test/test_rename.py
@@ -1,0 +1,8 @@
+import unittest
+from features import SFTP
+from unittest.mock import patch
+
+
+class UnitTests(unittest.TestCase):
+    def test_rename(self):
+        pass
